% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxforR.R
\name{oxforr_country}
\alias{oxforr_country}
\title{oxforr_country}
\usage{
oxforr_country(country)
}
\arguments{
\item{country}{Country's ISO code.}
}
\value{

}
\description{
This function allows you to find and search the right country code associated with the Oxford COVID-19 Government Response Tracker (OxCGRT).
If no argument is filed, all indicators will be displayed.
}
