% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxforR.R
\name{oxforr_data}
\alias{oxforr_data}
\title{Title}
\usage{
oxforr_data(country = data_long_country, indicator = data_long_indicator)
}
\arguments{
\item{country}{Countries' ISO code.}

\item{indicator}{Indicators from the Oxford COVID-19 Government Response Tracker (OxCGRT)}
}
\value{
Data for the country and indicator requested
}
\description{
This function allows you to find and display the Oxford COVID-19 Government Response Tracker (OxCGRT) data according to the selected parameters.
If no arguments are filled, all data will be displayed.
}
\examples{
myData <-oxforr_data("CAN","C1")
}
