% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxforR.R
\name{oxforr_indicator}
\alias{oxforr_indicator}
\title{oxforr_indicator}
\usage{
oxforr_indicator(indicators)
}
\arguments{
\item{indicators}{Indicator code}
}
\value{
list of indicators
}
\description{
This function allows you to find and search the right indicator code from the  The Oxford COVID-19 Government Response Tracker (OxCGRT) you want to use.
If no argument is filed, all indicators will be displayed.
}
\examples{
indicators <-oxforr_indicator()
public_indicators <-oxforr_indicator("public")
}
